requirements_dev.txt
-r requirements.txt
codecov==2.1.12
coloredlogs==15.0.1
coverage==6.5.0
django-naomi==0.8
factory_boy==3.2.1
flake8-blind-except==0.2.1
flake8-commas==2.1.0
flake8-docstrings==1.6.0
flake8-isort==5.0.3
flake8-pep3101==2.0.0
flake8-quotes==3.3.1
flake8-string-format==0.3.0
flake8==5.0.4
ipython==8.7.0
isort==5.10.1
mock==4.0.3
pydocstyle==6.1.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytest==7.2.0
python-coveralls==2.9.3
sphinx==5.3.0
django-debug-toolbar==3.8.1
channels==4.0.0
channels_redis==4.1.0